requirements.txt
# Core RAG dependencies
chromadb>=1.3.4
ollama>=0.4.0
rank-bm25>=0.2.2
tqdm>=4.67.1

# Document parsers
python-docx>=1.2.0
PyPDF2>=3.0.1

# LangChain text splitter (used by PDF parser)
langchain-text-splitters>=0.3.0

# Utilities
python-dotenv>=1.2.1
